Restructure model, routes, and curl-scripts folders
Change idea of a cart model to purchase model, rearrange routes and
curl-scripts of cart and cartProduct into unused folders to reflect new
purchase model structure. Simplify product model, rearrange productCart
and productGenereate files and associated routes and curl scripts to
designated unused folders. Add new product and purchase routes to
server.js.
