+++ app/models/product.js
@@ -12,13 +12,9 @@ const productSchema = new mongoose.Schema({
   price: {
     type: Number,
     required: true
-  },
-  productCount: {
-    type: Number,
-    min: 1
   }
 }, {
   timestamps: true
 })
 
-module.exports = productSchema
+module.exports = mongoose.model('Product', productSchema)
