+++ app/models/purchase.js
@@ -1,0 +1,18 @@
+const mongoose = require('mongoose')
+
+const purchaseSchema = new mongoose.Schema({
+  product: {
+    type: mongoose.Schema.Types.ObjectId,
+    ref: 'Product',
+    required: true
+  },
+  owner: {
+    type: mongoose.Schema.Types.ObjectId,
+    ref: 'User',
+    required: true
+  }
+}, {
+  timestamps: true
+})
+
+module.exports = mongoose.model('Purchase', purchaseSchema)
