+++ app/models/unused-models/productGenerate.js
@@ -1,0 +1,25 @@
+// Unused resource
+// const mongoose = require('mongoose')
+//
+// const productGenerateSchema = new mongoose.Schema({
+//   name: {
+//     type: String,
+//     required: true
+//   },
+//   description: {
+//     type: String,
+//     required: true
+//   },
+//   price: {
+//     type: Number,
+//     required: true
+//   },
+//   productCount: {
+//     type: Number,
+//     min: 1
+//   }
+// }, {
+//   timestamps: true
+// })
+//
+// module.exports = mongoose.model('Product', productGenerateSchema)
