+++ curl-scripts/unused-curl-scripts/cartProducts/create.sh
