+++ curl-scripts/unused-curl-scripts/cartProducts/destroy.sh
