+++ curl-scripts/unused-curl-scripts/carts/destroy.sh
