+++ curl-scripts/unused-curl-scripts/carts/show.sh
