+++ server.js
@@ -6,9 +6,12 @@ const cors = require('cors')
 // require route files
 const exampleRoutes = require('./app/routes/example_routes')
 const userRoutes = require('./app/routes/user_routes')
-const productGenerateRoutes = require('./app/routes/productGenerate_routes')
-const cartRoutes = require('./app/routes/cart_routes')
-const cartProductRoutes = require('./app/routes/cartProduct_routes')
+const productRoutes = require('./app/routes/product_routes')
+const purchaseRoutes = require('./app/routes/purchase_routes')
+// const productGenerateRoutes = require('./app/routes/productGenerate_routes')
+// const cartRoutes = require('./app/routes/cart_routes')
+// const cartProductRoutes = require('./app/routes/cartProduct_routes')
+
 // require middleware
 const errorHandler = require('./lib/error_handler')
 const requestLogger = require('./lib/request_logger')
@@ -59,9 +62,11 @@ app.use(requestLogger)
 // register route files
 app.use(exampleRoutes)
 app.use(userRoutes)
-app.use(productGenerateRoutes)
-app.use(cartRoutes)
-app.use(cartProductRoutes)
+app.use(productRoutes)
+app.use(purchaseRoutes)
+// app.use(productGenerateRoutes)
+// app.use(cartRoutes)
+// app.use(cartProductRoutes)
 
 // register error handling middleware
 // note that this comes after the route middlewares, because it needs to be
